refactor(helpers): split helpers.rs into config/paths/names submodules
Replaces the single 760-LOC `src-tauri/src/helpers.rs` with a
`helpers/` directory of three focused submodules + a re-export
`mod.rs`. The flat `helpers::*` public surface every external caller
depends on (`aethon_dir`, `parse_config_toml`, `resolve_inside_root`,
`sanitize_filename_segment`, `validate_state_name`,
`normalize_default_share_mode`, `normalize_new_tab_kind`,
`FONT_SIZE_*`, `clamp_font_size`) is preserved.

- `config.rs` — TOML schema (`UiConfig`, `AgentConfig`, `ShellConfig`,
  `ShortcutsConfig`, `ExtensionsConfig`, `AethonConfig`),
  `parse_config_toml`, normalizers, `EXT_STATE_*` /
  `resolve_ext_state_limits`, `FONT_SIZE_*` / `clamp_font_size`, plus
  all parser/clamp tests.
- `paths.rs` — `aethon_dir` (with the `AETHON_USER_DIR` sandbox
  override) and `resolve_inside_root`, plus their tests (including
  the env-mutex `aethon_dir_tests` submodule for serialized env
  manipulation).
- `names.rs` — `validate_state_name`, `sanitize_filename_segment`,
  plus tests.

Inter-module dependency surface is zero; each submodule is
self-contained. The unused struct types (`AethonConfig`, etc.) stay
accessible via `helpers::config::*` for tests; the flat re-export
in `mod.rs` keeps only the names external callers actually use to
avoid dead-code warnings.
