+++ src-tauri/src/helpers/config.rs
@@ -1,47 +1,13 @@
-//! Pure helpers extracted from `lib.rs` so they're testable without a Tauri
-//! `AppHandle` or filesystem. Each function here does one well-defined
-//! thing and returns a deterministic result for a given input.
+//! Config schemas and parsers for `~/.aethon/config.toml`. The structs
+//! mirror the user-facing TOML keys; `parse_config_toml` flattens them
+//! into the canonical JSON shape the frontend consumes, with all
+//! defaults applied and unknown values clamped.
 //!
-//! Cargo's unit tests live at the bottom of this file (`#[cfg(test)]`).
+//! Also home to the small numeric clamps the config feeds: the
+//! extension state-size limits (`EXT_STATE_*`) and the UI font-size
+//! range. These constants live next to the parser that reads them.
 
 use serde::Deserialize;
-use std::path::PathBuf;
-
-/// Resolve the Aethon user directory. Honors the `AETHON_USER_DIR`
-/// environment variable when set (used by `scripts/dev.sh --new` to
-/// route a session into a per-PID tmp sandbox so first-run UX can be
-/// exercised without nuking the real user data). Falls back to
-/// `<home>/.aethon` otherwise. Caller is responsible for `home_dir`
-/// when no override is set — pass `None` to skip the fallback and get
-/// a `None` back when neither the env var nor a usable home is set.
-pub fn aethon_dir(home: Option<PathBuf>) -> Option<PathBuf> {
-    if let Ok(s) = std::env::var("AETHON_USER_DIR")
-        && !s.is_empty()
-    {
-        return Some(PathBuf::from(s));
-    }
-    home.map(|h| h.join(".aethon"))
-}
-
-/// Validates a leaf filename used inside `~/.aethon/`. Rejects anything
-/// that could escape the directory — empty, slashes, parent-directory
-/// references. Used by `read_state` / `write_state` to keep arbitrary
-/// callers from writing outside the user dir.
-pub fn validate_state_name(name: &str) -> Result<(), String> {
-    if name.is_empty() {
-        return Err("invalid state name: (empty)".to_string());
-    }
-    if name.contains('/') || name.contains('\\') {
-        return Err(format!("invalid state name: {name}"));
-    }
-    if name == ".." || name == "." {
-        return Err(format!("invalid state name: {name}"));
-    }
-    if name.starts_with('\0') || name.contains('\0') {
-        return Err(format!("invalid state name: {name}"));
-    }
-    Ok(())
-}
 
 /// Schema for `~/.aethon/config.toml`. Mirrors the `AethonConfig` family
 /// in `lib.rs` so the parsing tests can reuse the same shape without
@@ -246,128 +212,9 @@ pub fn clamp_font_size(size: u32) -> u32 {
     size.clamp(FONT_SIZE_MIN, FONT_SIZE_MAX)
 }
 
-/// Lexically resolve `..` and `.` segments in `path` and check whether
-/// the result is `root` or a descendant. Inputs must be absolute. Returns
-/// `Some(resolved)` when the path stays inside `root`, `None` otherwise.
-///
-/// This is the gatekeeper for the file-system Tauri commands in
-/// [`crate::commands::fs`]. Each editor / file-tree operation passes the
-/// active project's cwd as `root` and the target path as `path`; the
-/// command refuses to touch anything that lexically escapes the root.
-///
-/// Implementation notes:
-///
-/// - Pure path arithmetic. We do **not** call `canonicalize` here — for
-///   create operations the target path doesn't exist yet, and the helper
-///   has to give a stable answer either way. Symlink-aware canonicalization
-///   happens once per command, after this check passes, on whichever
-///   parent component already exists. That second pass catches symlink
-///   escapes that lexical resolution can't see.
-/// - Strips `RootDir`/`Prefix` components on Windows so the comparison is
-///   structural; the inputs are still required to be absolute.
-/// - Both arguments must be normalized to the same prefix style by the
-///   caller (the commands convert `tilde` and relative segments before
-///   calling).
-pub fn resolve_inside_root(
-    root: &std::path::Path,
-    path: &std::path::Path,
-) -> Option<std::path::PathBuf> {
-    use std::path::{Component, PathBuf};
-    if !root.is_absolute() || !path.is_absolute() {
-        return None;
-    }
-    // Walk `path` lexically; build up the resolved absolute path.
-    let mut resolved: Vec<Component<'_>> = Vec::with_capacity(8);
-    for component in path.components() {
-        match component {
-            Component::ParentDir => {
-                // Pop the most recent Normal component. If the only thing
-                // left is the prefix/root, this attempts to ascend past
-                // the filesystem root — refuse.
-                if let Some(last) = resolved.last()
-                    && matches!(last, Component::Normal(_))
-                {
-                    resolved.pop();
-                    continue;
-                }
-                return None;
-            }
-            Component::CurDir => continue,
-            other => resolved.push(other),
-        }
-    }
-    let resolved_path: PathBuf = resolved.iter().collect();
-    // Same-prefix structural compare. `starts_with` matches Path
-    // component-by-component, so `/a/bc` does not start with `/a/b`.
-    if resolved_path == root || resolved_path.starts_with(root) {
-        Some(resolved_path)
-    } else {
-        None
-    }
-}
-
-/// POSIX-friendly filename sanitiser: keeps `[A-Za-z0-9_-]+`, replaces
-/// runs of unsafe chars with `_`, trims leading/trailing dots/dashes/
-/// underscores, and clamps to 64 chars. Empty input → empty output (the
-/// caller substitutes a default stem).
-pub fn sanitize_filename_segment(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let mut last_was_underscore = false;
-    for c in input.chars() {
-        let ok = c.is_ascii_alphanumeric() || c == '-' || c == '_';
-        if ok {
-            out.push(c);
-            last_was_underscore = false;
-        } else if !last_was_underscore && !out.is_empty() {
-            out.push('_');
-            last_was_underscore = true;
-        }
-    }
-    let trimmed = out
-        .trim_matches(|c: char| c == '_' || c == '-' || c == '.')
-        .to_string();
-    if trimmed.len() > 64 {
-        trimmed.chars().take(64).collect()
-    } else {
-        trimmed
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // ── validate_state_name ────────────────────────────────────────────────
-    #[test]
-    fn validate_state_name_accepts_simple_leaves() {
-        assert!(validate_state_name("messages.json").is_ok());
-        assert!(validate_state_name("theme").is_ok());
-        assert!(validate_state_name("config.toml").is_ok());
-        assert!(validate_state_name("state.json").is_ok());
-    }
-
-    #[test]
-    fn validate_state_name_rejects_empty() {
-        assert!(validate_state_name("").is_err());
-    }
-
-    #[test]
-    fn validate_state_name_rejects_path_separators() {
-        assert!(validate_state_name("a/b").is_err());
-        assert!(validate_state_name("a\\b").is_err());
-        assert!(validate_state_name("/absolute").is_err());
-    }
-
-    #[test]
-    fn validate_state_name_rejects_dot_traversal() {
-        assert!(validate_state_name("..").is_err());
-        assert!(validate_state_name(".").is_err());
-    }
-
-    #[test]
-    fn validate_state_name_rejects_null_bytes() {
-        assert!(validate_state_name("foo\0bar").is_err());
-    }
 
     // ── parse_config_toml ──────────────────────────────────────────────────
     #[test]
@@ -639,122 +486,5 @@ state_hard_kb = 100
     fn clamp_font_size_clamps_high() {
         assert_eq!(clamp_font_size(99), FONT_SIZE_MAX);
         assert_eq!(clamp_font_size(FONT_SIZE_MAX + 1), FONT_SIZE_MAX);
-    }
-
-    // ── resolve_inside_root ────────────────────────────────────────────────
-    use std::path::{Path, PathBuf};
-
-    #[test]
-    fn resolve_inside_root_accepts_direct_descendant() {
-        let root = Path::new("/projects/aethon");
-        let target = Path::new("/projects/aethon/src/App.tsx");
-        let out = resolve_inside_root(root, target).expect("should resolve");
-        assert_eq!(out, PathBuf::from("/projects/aethon/src/App.tsx"));
-    }
-
-    #[test]
-    fn resolve_inside_root_accepts_root_itself() {
-        let root = Path::new("/projects/aethon");
-        let out = resolve_inside_root(root, root).expect("root is inside root");
-        assert_eq!(out, PathBuf::from("/projects/aethon"));
-    }
-
-    #[test]
-    fn resolve_inside_root_resolves_inner_parent_segments() {
-        // /projects/aethon/src/.. → /projects/aethon, still inside.
-        let root = Path::new("/projects/aethon");
-        let target = Path::new("/projects/aethon/src/..");
-        let out = resolve_inside_root(root, target).expect("inner .. stays inside");
-        assert_eq!(out, PathBuf::from("/projects/aethon"));
-    }
-
-    #[test]
-    fn resolve_inside_root_rejects_traversal_escape() {
-        let root = Path::new("/projects/aethon");
-        // /projects/aethon/../passwd → /projects/passwd, escapes.
-        assert!(resolve_inside_root(root, Path::new("/projects/aethon/../passwd")).is_none());
-        // /etc/passwd is plain outside.
-        assert!(resolve_inside_root(root, Path::new("/etc/passwd")).is_none());
-        // Sibling that shares a prefix is NOT a descendant — `/projects/aethon-other`
-        // starts with the same string as `/projects/aethon` but is a sibling
-        // dir. starts_with compares components, so this is rejected.
-        assert!(resolve_inside_root(root, Path::new("/projects/aethon-other/file")).is_none());
-    }
-
-    #[test]
-    fn resolve_inside_root_rejects_relative_inputs() {
-        // Both args must be absolute. A relative root or path is a caller
-        // bug — return None so the command surfaces an error.
-        assert!(resolve_inside_root(Path::new("projects/aethon"), Path::new("/x")).is_none());
-        assert!(resolve_inside_root(Path::new("/x"), Path::new("projects/aethon")).is_none());
-    }
-
-    #[test]
-    fn resolve_inside_root_rejects_pop_past_root() {
-        // /projects/aethon/../.. ascends above /projects — refuse rather
-        // than pop off the prefix component.
-        let root = Path::new("/projects/aethon");
-        assert!(resolve_inside_root(root, Path::new("/projects/aethon/../..")).is_none());
-    }
-
-    /// Tests for `aethon_dir`. The function consults `AETHON_USER_DIR`
-    /// at call time, so each test must set+unset the var locally to
-    /// stay isolated from sibling tests running in the same process.
-    /// We use a global mutex (`ENV_LOCK`) so concurrent test threads
-    /// can't observe each other's env mutations.
-    mod aethon_dir_tests {
-        use super::super::aethon_dir;
-        use std::path::PathBuf;
-        use std::sync::Mutex;
-
-        static ENV_LOCK: Mutex<()> = Mutex::new(());
-
-        #[test]
-        fn returns_home_dotaethon_when_no_override() {
-            let _g = ENV_LOCK.lock().unwrap();
-            // SAFETY: ENV_LOCK serialises env mutations across tests in
-            // this module so concurrent test threads cannot observe a
-            // half-written global.
-            unsafe { std::env::remove_var("AETHON_USER_DIR") };
-            let got = aethon_dir(Some(PathBuf::from("/home/test")));
-            assert_eq!(got, Some(PathBuf::from("/home/test/.aethon")));
-        }
-
-        #[test]
-        fn returns_override_when_env_set() {
-            let _g = ENV_LOCK.lock().unwrap();
-            unsafe { std::env::set_var("AETHON_USER_DIR", "/tmp/sandbox-42") };
-            let got = aethon_dir(Some(PathBuf::from("/home/test")));
-            assert_eq!(got, Some(PathBuf::from("/tmp/sandbox-42")));
-            unsafe { std::env::remove_var("AETHON_USER_DIR") };
-        }
-
-        #[test]
-        fn returns_none_when_no_home_and_no_env() {
-            let _g = ENV_LOCK.lock().unwrap();
-            unsafe { std::env::remove_var("AETHON_USER_DIR") };
-            assert_eq!(aethon_dir(None), None);
-        }
-
-        #[test]
-        fn env_override_wins_even_when_no_home() {
-            let _g = ENV_LOCK.lock().unwrap();
-            unsafe { std::env::set_var("AETHON_USER_DIR", "/tmp/sandbox-99") };
-            let got = aethon_dir(None);
-            assert_eq!(got, Some(PathBuf::from("/tmp/sandbox-99")));
-            unsafe { std::env::remove_var("AETHON_USER_DIR") };
-        }
-
-        #[test]
-        fn empty_env_var_falls_back_to_home() {
-            let _g = ENV_LOCK.lock().unwrap();
-            // Some shells export empty string instead of unsetting. We
-            // treat that as "no override" so the user isn't trapped in
-            // a broken sandbox at "".
-            unsafe { std::env::set_var("AETHON_USER_DIR", "") };
-            let got = aethon_dir(Some(PathBuf::from("/h")));
-            assert_eq!(got, Some(PathBuf::from("/h/.aethon")));
-            unsafe { std::env::remove_var("AETHON_USER_DIR") };
-        }
     }
 }
